requirements.txt
Django==3.2.2
python-dotenv==0.17.1
django-environ~=0.4.5
djangorestframework~=3.12.4
pytz~=2021.1
django-debug-toolbar~=3.2.1
django-braces~=1.14.0
python-dateutil~=2.8.1